% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data}
\usage{
normalize_data(data)
}
\arguments{
\item{data}{A data frame to be normalized.}
}
\value{
A normalized data frame.
}
\description{
This function normalizes the data by dividing each element in columns 2 to the last column by the sum of that column.
}
\examples{
data <- data.frame(ID = 1:5, A = c(10, 20, 30, 40, 50), B = c(5, 10, 15, 20, 25))
normalized_data <- normalize_data(data)
}
