% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIKOR.R
\name{VIKOR}
\alias{VIKOR}
\title{Calculate VIKOR rankings}
\usage{
VIKOR(matrix, weights, q, v)
}
\arguments{
\item{matrix}{Decision matrix where rows represent alternatives and columns represent criteria.}

\item{weights}{Criteria weights.}

\item{q}{Criteria direction ("min" for minimization, "max" for maximization).}

\item{v}{Parameter (0 <= v <= 1).}
}
\value{
A vector of rankings.
}
\description{
This function implements the VIKOR algorithm to rank alternatives based on multiple criteria.
}
\examples{
decision_matrix <- matrix(c(3, 5, 7, 4, 6, 8, 2, 7, 9, 6, 5, 8), nrow = 4, byrow = TRUE)
weights <- c(0.4, 0.6)
direction <- c("min", "max")
v <- 0.5
VIKOR(decision_matrix, weights, direction, v)

}
